Split constants and config
