+++ ui/react-components/api/providerApi.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import mockAxios from 'jest-mock-axios';
 import {getAllProviders} from "./providerApi";
-import {providerUrl, providerParams} from "../constants";
+import {providerUrl, providerParams} from "../config";
 
 afterEach(() => {
     mockAxios.reset();
