+++ ui/react-components/components/Location/LocationSearch.jsx
@@ -3,7 +3,7 @@ import React, {useEffect, useState} from "react";
 import PropTypes from "prop-types";
 import {injectIntl} from 'react-intl';
 import {getAllByTag} from "../../api/locationApi";
-import {locationTagName} from "../../constants";
+import {locationTagName} from "../../config";
 import {forEach} from 'lodash';
 
 const LocationSearch = (props) => {
